refactor: addres PR review feedback
